requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
coverage==7.5.4
cryptography==42.0.8
decorator==5.1.1
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.6
django-extensions==3.2.3
django-ninja==1.2.0
executing==2.0.1
filelock==3.15.4
gunicorn==22.0.0
h11==0.14.0
identify==2.5.36
ipython==8.25.0
jedi==0.19.1
matplotlib-inline==0.1.7
nodeenv==1.9.1
packaging==24.1
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
pre-commit==3.7.1
prompt_toolkit==3.0.47
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
python-dotenv==1.0.1
PyYAML==6.0.1
ruff==0.5.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
uvicorn==0.30.1
virtualenv==20.26.3
wcwidth==0.2.13
whitenoise==6.7.0
